// openapi-v2/models/APICreateCustomerPayload.ts
/**
 * 24k
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class APICreateCustomerPayload {
    /**
    * Name
    */
    'name': string;
    /**
    * External ID
    */
    'externalId': string;
    /**
    * Construct a type with a set of properties K of type T
    */
    'metadata'?: { [key: string]: any; };
    /**
    * Aliases
    */
    'aliases'?: Array<string>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "externalId",
            "baseName": "externalId",
            "type": "string",
            "format": ""
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "{ [key: string]: any; }",
            "format": ""
        },
        {
            "name": "aliases",
            "baseName": "aliases",
            "type": "Array<string>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return APICreateCustomerPayload.attributeTypeMap;
    }

    public constructor() {
    }
}