// openapi-v2/models/APICustomer.ts
/**
 * 24k
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* API Customer
*/
export class APICustomer {
    /**
    * ID
    */
    'id': string;
    /**
    * Name
    */
    'name': string;
    /**
    * Account ID
    */
    'accountId': string;
    /**
    * Aliases
    */
    'aliases': Array<string>;
    /**
    * External ID
    */
    'externalId': string;
    /**
    * Construct a type with a set of properties K of type T
    */
    'metadata': { [key: string]: any; };
    /**
    * Created At
    */
    'createdAt': string;
    /**
    * Updated At
    */
    'updatedAt': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "accountId",
            "baseName": "accountId",
            "type": "string",
            "format": ""
        },
        {
            "name": "aliases",
            "baseName": "aliases",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "externalId",
            "baseName": "externalId",
            "type": "string",
            "format": ""
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "{ [key: string]: any; }",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "createdAt",
            "type": "string",
            "format": "date"
        },
        {
            "name": "updatedAt",
            "baseName": "updatedAt",
            "type": "string",
            "format": "date"
        }    ];

    static getAttributeTypeMap() {
        return APICustomer.attributeTypeMap;
    }

    public constructor() {
    }
}