// openapi-v2/models/APIEvent.ts
/**
 * 24k
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* APIEvent model
*/
export class APIEvent {
    /**
    * ID
    */
    'id': string;
    /**
    * Name
    */
    'name': string;
    /**
    * Timestamp - When the event occurred
    */
    'timestamp': number;
    /**
    * Construct a type with a set of properties K of type T
    */
    'metadata': { [key: string]: any; };
    /**
    * Account ID
    */
    'accountId': string;
    /**
    * Customer ID
    */
    'customerId': string;
    /**
    * Product ID
    */
    'productId': string;
    /**
    * Reference - unique identifier for the event
    */
    'ref': string;
    /**
    * Created At
    */
    'createdAt': string;
    /**
    * Updated At
    */
    'updatedAt': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "timestamp",
            "baseName": "timestamp",
            "type": "number",
            "format": "double"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "{ [key: string]: any; }",
            "format": ""
        },
        {
            "name": "accountId",
            "baseName": "accountId",
            "type": "string",
            "format": ""
        },
        {
            "name": "customerId",
            "baseName": "customerId",
            "type": "string",
            "format": ""
        },
        {
            "name": "productId",
            "baseName": "productId",
            "type": "string",
            "format": ""
        },
        {
            "name": "ref",
            "baseName": "ref",
            "type": "string",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "createdAt",
            "type": "string",
            "format": "date"
        },
        {
            "name": "updatedAt",
            "baseName": "updatedAt",
            "type": "string",
            "format": "date"
        }    ];

    static getAttributeTypeMap() {
        return APIEvent.attributeTypeMap;
    }

    public constructor() {
    }
}