// openapi-v2/models/APIInvoiceBillingPeriod.ts
/**
 * 24k
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Billing Period
*/
export class APIInvoiceBillingPeriod {
    /**
    * Period End Date
    */
    'endTime': string;
    /**
    * Period Start Date
    */
    'startTime': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "endTime",
            "baseName": "endTime",
            "type": "string",
            "format": "date"
        },
        {
            "name": "startTime",
            "baseName": "startTime",
            "type": "string",
            "format": "date"
        }    ];

    static getAttributeTypeMap() {
        return APIInvoiceBillingPeriod.attributeTypeMap;
    }

    public constructor() {
    }
}