// openapi-v2/models/APIQueryPayloadAPIEvent.ts
/**
 * 24k
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class APIQueryPayloadAPIEvent {
    /**
    * Filter by field.
    */
    'filter': any;
    /**
    * Page
    */
    'page'?: number;
    /**
    * Page Size
    */
    'pageSize'?: number;
    /**
    * Sort by field. Use 1 for ascending and -1 for descending.
    */
    'sort'?: any;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "filter",
            "baseName": "filter",
            "type": "any",
            "format": ""
        },
        {
            "name": "page",
            "baseName": "page",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "pageSize",
            "baseName": "pageSize",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "sort",
            "baseName": "sort",
            "type": "any",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return APIQueryPayloadAPIEvent.attributeTypeMap;
    }

    public constructor() {
    }
}